[post] update: update/delete post API

+++ Models/DTO/Post/CreatePostDTO.cs
@@ -4,11 +4,11 @@ public class CreatePostDTO
 {
     public string Title { get; set; }
     public string Description { get; set; }
-    public double Area { get; set; }
+    public float Area { get; set; }
     public string Address { get; set; }
     public int AddressWardId { get; set; }
-    public double Price { get; set; }
-    public double PrePaidPrice { get; set; }
+    public float Price { get; set; }
+    public float PrePaidPrice { get; set; }
     public int CategoryId { get; set; }
     public int LimitTenant { get; set; }
     public List<CreateMediaDTO> Medias { get; set; }
